Set NBD max_sectors_kb to 128 to align with bmcweb buffer size
The bmcweb hard-codes the NBD buffer size to 131088 bytes
(256 * 512 + 16, where 16 is the NBD header).
This means the kernel's max_user_sectors must be set to 256
and max_sectors_kb must be set to 128 to match bmcweb's expectation.

See:
```
https://gitlab-master.nvidia.com/dgx/bmc/bmcweb/-/blob/develop/include/nbd_proxy.hpp?ref_type=heads#L47
https://gitlab-master.nvidia.com/dgx/bmc/bmcweb/-/blob/develop/http/websocket.hpp?ref_type=heads#L90
```

In Linux kernel 6.5.12, the default max_sectors_kb was already 128,
so the issue didn't appear.
In Linux kernel 6.12.9, the default changed to 1280, which causes
the mismatch and triggers the failure.

So we can echo to sysfs directly:

echo 128 > /sys/block/nbd0/queue/max_sectors_kb

This avoids modifying the kernel and aligns with the buffer size
expected by bmcweb.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5447454

Signed-off-by: Shane Lin <hslin@nvidia.com>